+++ npm-package.js
@@ -5,12 +5,16 @@
 */
 
 /*  external requirements  */
-const os    = require("os")
-const path  = require("path")
-const glob  = require("glob")
-const shell = require("shelljs")
-const execa = require("execa")
-const zip   = require("cross-zip")
+const fs        = require("fs")
+const os        = require("os")
+const path      = require("path")
+const glob      = require("glob")
+const shell     = require("shelljs")
+const execa     = require("execa")
+const zip       = require("cross-zip")
+const DSIG      = require("dsig")
+const Prompt    = require("prompt-base")
+const PromptPW  = require("prompt-password")
 
 /*  establish asynchronous environment  */
 ;(async () => {
@@ -26,6 +30,25 @@ const zip   = require("cross-zip")
     for (const file of remove)
         shell.rm("-rf", file)
 
+    /*  helper function for digitally signing distribution artifact  */
+    const sign = async (zipfile) => {
+        console.log("++ generating digital signature for ZIP distribution archive")
+        const prompt = new PromptPW({
+            type:    "password",
+            message: "Password",
+            name:    "password"
+        })
+        const passPhrase = await prompt.run()
+        const sigfile = zipfile.replace(/\.zip$/, ".sig")
+        const payload = await fs.promises.readFile(zipfile, { encoding: null })
+        const privateKey = await fs.promises.readFile(
+            path.join(os.homedir(), ".dsig", "LiVE.prv"), { encoding: "utf8" })
+        const signature = await DSIG.sign(payload, privateKey, passPhrase)
+        await fs.promises.writeFile(sigfile, signature, { encoding: "utf8" })
+        const publicKey = await fs.promises.readFile("npm-package.pk", { encoding: "utf8" })
+        await DSIG.verify(payload, signature, publicKey)
+    }
+
     /*   package according to platform...  */
     const electronbuilder = path.resolve(path.join("node_modules", ".bin", "electron-builder"))
     if (os.platform() === "win32") {
@@ -39,8 +62,8 @@ const zip   = require("cross-zip")
         console.log("++ packing App into ZIP distribution archive")
         zip.zipSync(
             path.join(__dirname, "dist/LiVE-Receiver.exe"),
-            path.join(__dirname, "dist/LiVE-Receiver-win-x64.zip")
-        )
+            path.join(__dirname, "dist/LiVE-Receiver-win-x64.zip"))
+        await sign("dist/LiVE-Receiver-win-x64.zip")
     }
     else if (os.platform() === "darwin") {
         /*  run Electron-Builder to package the application  */
@@ -54,8 +77,8 @@ const zip   = require("cross-zip")
         shell.mv("dist/mac/LiVE-Receiver.app", "dist/LiVE-Receiver.app")
         zip.zipSync(
             path.join(__dirname, "dist/LiVE-Receiver.app"),
-            path.join(__dirname, "dist/LiVE-Receiver-mac-x64.zip")
-        )
+            path.join(__dirname, "dist/LiVE-Receiver-mac-x64.zip"))
+        await sign("dist/LiVE-Receiver-mac-x64.zip")
     }
     else if (os.platform() === "linux") {
         /*  run Electron-Builder to package the application  */
@@ -68,8 +91,8 @@ const zip   = require("cross-zip")
         shell.mv("dist/LiVE-Receiver-*.AppImage", "dist/LiVE-Receiver")
         zip.zipSync(
             path.join(__dirname, "dist/LiVE-Receiver"),
-            path.join(__dirname, "dist/LiVE-Receiver-lnx-x64.zip")
-        )
+            path.join(__dirname, "dist/LiVE-Receiver-lnx-x64.zip"))
+        await sign("dist/LiVE-Receiver-lnx-x64.zip")
     }
 })().catch((err) => {
     console.log(`** package: ERROR: ${err}`)